backend/requirements.txt
fastapi==0.111.0
sqlalchemy==2.0.30
psycopg2-binary==2.9.9
sse-starlette==1.8.2
uvicorn[standard]==0.30.1
python-multipart==0.0.9
google-generativeai==0.7.0
pydantic-settings==2.3.1
pymupdf4llm==0.0.10
python-dotenv==1.0.1
aiohttp==3.9.1
pytest==7.4.3
pytest-asyncio==0.23.2
pytest-cov==5.0.0
ruff==0.4.4
pandas==2.2.2
tabulate==0.9.0
hvac==2.3.0
